# Help the compiler vectorize `adjacent_difference`