--- a/stl/inc/numeric
+++ b/stl/inc/numeric
@@ -462,7 +462,7 @@ _FwdIt2 transform_inclusive_scan(_ExPo&& _Exec, _FwdIt1 _First, _FwdIt1 _Last, _
 #endif // _HAS_CXX17
 
 template <class _Ty>
-_CONSTEXPR20 bool _Arrays_overlap(const _Ty* const _First, const _Ty* const _Second, const size_t _Common_count) {
+bool _Arrays_overlap(const _Ty* const _First, const _Ty* const _Second, const size_t _Common_count) {
     const uintptr_t _First_val  = reinterpret_cast<uintptr_t>(_First);
     const uintptr_t _Second_val = reinterpret_cast<uintptr_t>(_Second);
     const size_t Size_bytes     = _Common_count * sizeof(_Ty);
@@ -470,7 +470,7 @@ _CONSTEXPR20 bool _Arrays_overlap(const _Ty* const _First, const _Ty* const _Sec
 }
 
 template <class _Ty, class _BinOp>
-_CONSTEXPR20 void _Adjacent_difference_no_overlap(
+void _Adjacent_difference_no_overlap(
     _Ty* __restrict _Dest, const _Ty* __restrict _Src, const size_t _Count, _BinOp _Func) {
     _Dest[0] = _Src[0];
     for (size_t _Ix = 1; _Ix != _Count; ++_Ix) {
@@ -488,19 +488,25 @@ _CONSTEXPR20 _OutIt adjacent_difference(const _InIt _First, const _InIt _Last, _
     if (_UFirst != _ULast) {
         if constexpr (_Iterators_are_contiguous<_InIt, _OutIt> && !_Iterator_is_volatile<_InIt>
                       && is_trivially_copyable_v<_Iter_value_t<_InIt>>) {
-            // Go with pointers and without loop-carried dependency to enable vectorization
-            const auto _Count    = _ULast - _UFirst;
-            const auto _Src_ptr  = _To_address(_UFirst);
-            const auto _Dest_ptr = _To_address(_Dest);
-            // Need to perform aliasing analysis.
-            // The vectorizer is generally able to do that on its own, and would guard the vectorized code with that,
-            // but when we elimniate loop-carried dependency we change the semantic of the unvectorized code too.
-            // So we need to perform this check manually, and after that can tell the compiler that there's no aliasing,
-            // to avoid it checking for that again.
-            if (!_Arrays_overlap(_Dest_ptr, _Src_ptr, static_cast<size_t>(_Count))) {
-                _Adjacent_difference_no_overlap(_Dest_ptr, _Src_ptr, static_cast<size_t>(_Count), _STD _Pass_fn(_Func));
-                _STD _Seek_wrapped(_Dest, _UDest + _Count);
-                return _Dest;
+#if _HAS_CXX20
+            if (!_STD is_constant_evaluated())
+#endif
+            {
+                // Go with pointers and without loop-carried dependency to enable vectorization
+                const auto _Count    = _ULast - _UFirst;
+                const auto _Src_ptr  = _To_address(_UFirst);
+                const auto _Dest_ptr = _To_address(_Dest);
+                // Need to perform aliasing analysis.
+                // The vectorizer is generally able to do that on its own, and would guard the vectorized code with
+                // that, but when we elimniate loop-carried dependency we change the semantic of the unvectorized code
+                // too. So we need to perform this check manually, and after that can tell the compiler that there's no
+                // aliasing, to avoid it checking for that again.
+                if (!_Arrays_overlap(_Dest_ptr, _Src_ptr, static_cast<size_t>(_Count))) {
+                    _Adjacent_difference_no_overlap(
+                        _Dest_ptr, _Src_ptr, static_cast<size_t>(_Count), _STD _Pass_fn(_Func));
+                    _STD _Seek_wrapped(_Dest, _UDest + _Count);
+                    return _Dest;
+                }
             }
         }