--- a/stl/inc/numeric
+++ b/stl/inc/numeric
@@ -461,14 +461,6 @@ _FwdIt2 transform_inclusive_scan(_ExPo&& _Exec, _FwdIt1 _First, _FwdIt1 _Last, _
     _UnaryOp _Transform_op) noexcept; // terminates
 #endif // _HAS_CXX17
 
-template <class _Ty1, class _Ty2>
-bool _Arrays_overlap(const _Ty1* const _First, const _Ty2* const _Second, const size_t _Common_count) {
-    const uintptr_t _First_val  = reinterpret_cast<uintptr_t>(_First);
-    const uintptr_t _Second_val = reinterpret_cast<uintptr_t>(_Second);
-    return _First_val + _Common_count * sizeof(_Ty1) > _Second_val
-        && _Second_val + _Common_count * sizeof(_Ty2) > _First_val;
-}
-
 template <class _TyDest, class _TySrc, class _BinOp>
 void _Adjacent_difference_no_overlap(
     _TyDest* const __restrict _Dest, _TySrc* const __restrict _Src, const size_t _Count, _BinOp _Func) {
@@ -498,17 +490,17 @@ _CONSTEXPR20 _OutIt adjacent_difference(const _InIt _First, const _InIt _Last, _
 #endif
             {
                 // Go with pointers and without loop-carried dependency to enable vectorization
-                const auto _Count    = _ULast - _UFirst;
-                const auto _Src_ptr  = _STD _To_address(_UFirst);
-                const auto _Dest_ptr = _STD _To_address(_Dest);
+                const auto _Count         = _ULast - _UFirst;
+                const auto _Count_as_size = static_cast<size_t>(_Count);
+                const auto _Src_ptr       = _STD _To_address(_UFirst);
+                const auto _Dest_ptr      = _STD _To_address(_Dest);
                 // Need to perform aliasing analysis.
                 // The vectorizer is generally able to do that on its own, and would guard the vectorized code with
                 // that, but when we elimniate loop-carried dependency we change the semantic of the unvectorized code
                 // too. So we need to perform this check manually, and after that can tell the compiler that there's no
                 // aliasing, to avoid it checking for that again.
-                if (!_STD _Arrays_overlap(_Dest_ptr, _Src_ptr, static_cast<size_t>(_Count))) {
-                    _STD _Adjacent_difference_no_overlap(
-                        _Dest_ptr, _Src_ptr, static_cast<size_t>(_Count), _STD _Pass_fn(_Func));
+                if (_Dest_ptr + _Count_as_size <= _Src_ptr || _Src_ptr + _Count_as_size <= _Dest_ptr) {
+                    _STD _Adjacent_difference_no_overlap(_Dest_ptr, _Src_ptr, _Count_as_size, _STD _Pass_fn(_Func));
                     _STD _Seek_wrapped(_Dest, _UDest + _Count);
                     return _Dest;
                 }