--- a/stl/inc/numeric
+++ b/stl/inc/numeric
@@ -496,7 +496,7 @@ _CONSTEXPR20 _OutIt adjacent_difference(const _InIt _First, const _InIt _Last, _
                 const auto _Dest_ptr      = _STD _To_address(_UDest);
                 // Need to perform aliasing analysis.
                 // The vectorizer is generally able to do that on its own, and would guard the vectorized code with
-                // that, but when we elimniate loop-carried dependency we change the semantic of the unvectorized code
+                // that, but when we eliminate loop-carried dependency we change the semantics of the unvectorized code
                 // too. So we need to perform this check manually, and after that can tell the compiler that there's no
                 // aliasing, to avoid it checking for that again.
                 if (_Dest_ptr + _Count_as_size <= _Src_ptr || _Src_ptr + _Count_as_size <= _Dest_ptr) {