@AlexGuteniev 2024-09-15
> * 8-bit case failed to vectorize for no reason (didn't look up if it is known compiler issue, or to be reported)

Reported DevCom-10745948