@CaseyCarter 2024-09-15
> > * 8-bit case failed to vectorize for no reason (didn't look up if it is known compiler issue, or to be reported)

Interestingly [it vectorizes if we use `-` directly instead of indirecting through `std::minus`](https://godbolt.org/z/4KMqx1rMa), [or if the output is a pointer to `int`](https://godbolt.org/z/rxccP83qf). Something to do with narrowing the result of the promoted operation, maybe?