@StephanTLavavej 2024-10-26
Thanks! :heart_eyes_cat: I pushed minor nitpicks and a significant fix for C++14/17. Speedups look good on my 5950X:

Benchmark           | Before |   After | Speedup
--------------------|-------:|--------:|-------:
`bm<uint8_t>/2255`  | 968 ns |  967 ns | 1.00
`bm<uint16_t>/2255` | 917 ns | 97.2 ns | 9.43
`bm<uint32_t>/2255` | 648 ns |  158 ns | 4.10
`bm<uint64_t>/2255` | 689 ns |  331 ns | 2.08
`bm<float>/2255`    | 646 ns |  158 ns | 4.09
`bm<double>/2255`   | 652 ns |  332 ns | 1.96